requirements.txt
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
certifi==2025.6.15
cffi==1.17.1
click==8.2.1
colorama==0.4.6
flake8==7.3.0
flake8-html==0.4.3
greenlet==3.2.3
iniconfig==2.1.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mccabe==0.7.0
packaging==25.0
pluggy==1.6.0
psycopg2-binary==2.9.10
pycodestyle==2.14.0
pycparser==2.22
pyflakes==3.4.0
Pygments==2.19.1
PyJWT==2.10.1
pytest==8.4.1
pytest-html==4.1.1
pytest-metadata==3.1.1
python-dotenv==1.1.0
sentry-sdk==2.32.0
SQLAlchemy==2.0.41
tabulate==0.9.0
typing_extensions==4.14.0
urllib3==2.5.0
